+++ botan-low/src/Botan/Low/SRP6.hs
@@ -8,19 +8,15 @@ Maintainer  : joris@well-typed.com, leo@apotheca.io
 Stability   : experimental
 Portability : POSIX
 
-The library contains an implementation of the SRP6-a password
-authenticated key exchange protocol.
-
+This module is based on the [Secure Remote
+Password](https://botan.randombit.net/handbook/api_ref/srp.html) section of the
+C++ API reference.
 -}
 
 module Botan.Low.SRP6 (
-
-  -- * Secure Random Password 6a
-  -- $introduction
-
-  -- * Usage
-  -- $usage
-
+    -- * Example usage
+    -- $usage
+    -- * Algorithm
     SRP6ServerSession(..)
   , withSRP6ServerSession
   , srp6ServerSessionInit
@@ -30,16 +26,12 @@ module Botan.Low.SRP6 (
   , srp6GenerateVerifier
   , srp6ClientAgree
   , srp6GroupSize
-
-  -- * SRP6 Types
-
+    -- * SRP6 Types
   , SRP6Verifier
   , SRP6BValue
   , SRP6AValue
   , SRP6SharedSecret
-
-  -- * SRP discrete logarithm groups
-
+    -- * SRP discrete logarithm groups
   , pattern MODP_SRP_1024
   , pattern MODP_SRP_1536
   , pattern MODP_SRP_2048
@@ -47,42 +39,18 @@ module Botan.Low.SRP6 (
   , pattern MODP_SRP_4096
   , pattern MODP_SRP_6144
   , pattern MODP_SRP_8192
-
   ) where
+
+import qualified Data.ByteString.Internal as BS
 
 import           Botan.Bindings.SRP6
 
 import           Botan.Low.Error
 import           Botan.Low.Hash
-import           Botan.Low.Make
 import           Botan.Low.Prelude
 import           Botan.Low.PubKey
 import           Botan.Low.Remake
 import           Botan.Low.RNG
-
-{- $introduction
-
-A SRP client provides what is called a SRP verifier to the server.
-This verifier is based on a password, but the password cannot be
-easily derived from the verifier (however brute force attacks are
-possible). Later, the client and server can perform an SRP exchange,
-which results in a shared secret key. This key can be used for
-mutual authentication and/or encryption.
-
-SRP works in a discrete logarithm group. Special parameter sets for
-SRP6 are defined, denoted in the library as @modp\/srp\/<size>@, for
-example @modp\/srp\/2048@.
-
-Warning
-
-While knowledge of the verifier does not easily allow an attacker to
-get the raw password, they could still use the verifier to impersonate
-the server to the client, so verifiers should be protected as carefully
-as a plaintext password would be.
-
-SRP6 may be used as part of SSL/TLS: https://www.rfc-editor.org/rfc/rfc5054
-
--}
 
 {- $usage
 
@@ -183,7 +151,7 @@ type SRP6SharedSecret = ByteString
 newtype SRP6ServerSession = MkSRP6ServerSession { getSRP6ServerSessionForeignPtr :: ForeignPtr BotanSRP6ServerSessionStruct }
 
 withSRP6ServerSession     :: SRP6ServerSession -> (BotanSRP6ServerSession -> IO a) -> IO a
--- | Destroy a SRP6 server session object immediately
+-- | Frees all resources of the SRP-6 server session object
 srp6ServerSessionDestroy  :: SRP6ServerSession -> IO ()
 createSRP6ServerSession   :: (Ptr BotanSRP6ServerSession -> IO CInt) -> IO SRP6ServerSession
 (_, withSRP6ServerSession, srp6ServerSessionDestroy, createSRP6ServerSession, _)
@@ -193,134 +161,199 @@ createSRP6ServerSession   :: (Ptr BotanSRP6ServerSession -> IO CInt) -> IO SRP6S
         botan_srp6_server_session_destroy
 
 -- | Initialize an SRP-6 server session object
-srp6ServerSessionInit
-    :: IO SRP6ServerSession -- ^ __srp6__: SRP-6 server session object
+srp6ServerSessionInit ::
+     IO SRP6ServerSession -- ^ __srp6__: SRP-6 server session object
 srp6ServerSessionInit = createSRP6ServerSession botan_srp6_server_session_init
 
--- | SRP-6 Server side step 1: Generate a server B-value
-srp6ServerSessionStep1
-    :: SRP6ServerSession    -- ^ __srp6__: SRP-6 server session object
-    -> SRP6Verifier         -- ^ __verifier[]__: the verification value saved from client registration
-    -> DLGroupName          -- ^ __group_id__: the SRP group id
-    -> HashName             -- ^ __hash_id__: the SRP hash in use
-    -> RNG                  -- ^ __rng_obj__: a random number generator object
-    -> IO SRP6BValue        -- ^ __B_pub[]__: out buffer to store the SRP-6 B value
-srp6ServerSessionStep1 srp6 verifier groupId hashId rng = withSRP6ServerSession srp6 $ \ srp6Ptr -> do
-    asBytesLen verifier $ \ verifierPtr verifierLen -> do
-        asCString groupId $ \ groupIdPtr -> do
-            asCString hashId $ \ hashIdPtr -> do
-                withRNG rng $ \ botanRNG -> do
-                    allocBytesQuerying $ \ outPtr outLen -> botan_srp6_server_session_step1
-                        srp6Ptr
-                        (ConstPtr verifierPtr)
-                        verifierLen
-                        (ConstPtr groupIdPtr)
-                        (ConstPtr hashIdPtr)
-                        botanRNG
-                        outPtr
-                        outLen
+-- | SRP-6 Server side step 1
+--
+-- NOTE: this function should be not be invoked twice on the same server
+-- session. Regardless of the result of the first invocation, the second
+-- invocation will result in an error. See
+-- https://github.com/randombit/botan/issues/5112 for more information. If a
+-- second invocation can not be prevented, try it on a newly initialised server
+-- session instead.
+srp6ServerSessionStep1 ::
+     SRP6ServerSession -- ^ __srp6__: SRP-6 server session object
+  -> SRP6Verifier      -- ^ __verifier[]__: the verification value saved from client registration
+  -> DLGroupName       -- ^ __group_id__: the SRP group id
+  -> HashName          -- ^ __hash_id__: the SRP hash in use
+  -> RNG               -- ^ __rng_obj__: a random number generator object
+  -> IO SRP6BValue     -- ^ __B_pub[]__: out buffer to store the SRP-6 B value
+srp6ServerSessionStep1 srp6 verifier groupId hashId rng =
+    withSRP6ServerSession srp6 $ \ srp6Ptr ->
+    asBytesLen verifier $ \ verifierPtr verifierLen ->
+    asCString groupId $ \ groupIdPtr ->
+    asCString hashId $ \ hashIdPtr ->
+    withRNG rng $ \ botanRNG ->
+    createWithGroupSize groupId $ \ outPtr outLen ->
+    throwBotanIfNegative_ $
+      botan_srp6_server_session_step1
+        srp6Ptr
+        (ConstPtr verifierPtr)
+        verifierLen
+        (ConstPtr groupIdPtr)
+        (ConstPtr hashIdPtr)
+        botanRNG
+        outPtr
+        outLen
 
--- | SRP-6 Server side step 2:  Generate the server shared key
-srp6ServerSessionStep2
-    :: SRP6ServerSession    -- ^ __srp6__: SRP-6 server session object
-    -> SRP6AValue           -- ^ __A[]__: the client's value
-    -> IO SRP6SharedSecret  -- ^ __key[]__: out buffer to store the symmetric key value
-srp6ServerSessionStep2 srp6 a = withSRP6ServerSession srp6 $ \ srp6Ptr -> do
-    asBytesLen a $ \ aPtr aLen -> do
-        allocBytesQuerying $ \ outPtr outLen -> botan_srp6_server_session_step2
-            srp6Ptr
-            (ConstPtr aPtr)
-            aLen
-            outPtr
-            outLen
+-- | SRP-6 Server side step 2
+srp6ServerSessionStep2 ::
+     SRP6ServerSession   -- ^ __srp6__: SRP-6 server session object
+  -> SRP6AValue          -- ^ __A[]__: the client's value
+  -> DLGroupName         -- ^ the SRP group id
+  -> IO SRP6SharedSecret -- ^ __key[]__: out buffer to store the symmetric key value
+srp6ServerSessionStep2 srp6 groupId a =
+    withSRP6ServerSession srp6 $ \ srp6Ptr ->
+    asBytesLen a $ \ aPtr aLen ->
+    createWithGroupSize groupId $ \ outPtr outLen ->
+    throwBotanIfNegative_ $
+      botan_srp6_server_session_step2
+        srp6Ptr
+        (ConstPtr aPtr)
+        aLen
+        outPtr
+        outLen
 
--- | SRP-6 Client side step 1:  Generate a new SRP-6 verifier
-srp6GenerateVerifier
-    :: Identifier       -- ^ __identifier__: a username or other client identifier
-    -> Password         -- ^ __password__: the secret used to authenticate user
-    -> Salt             -- ^ __salt[]__: a randomly chosen value, at least 128 bits long
-    -> DLGroupName      -- ^ __group_id__: specifies the shared SRP group
-    -> HashName         -- ^ __hash_id__: specifies a secure hash function
-    -> IO SRP6Verifier  -- ^ __verifier[]__: out buffer to store the SRP-6 verifier value
-srp6GenerateVerifier identifier password salt groupId hashId = asCString identifier $ \ identifierPtr -> do
-    asCString password $ \ passwordPtr -> do
-        asBytesLen salt $ \ saltPtr saltLen -> do
-            asCString groupId $ \ groupIdPtr -> do
-                asCString hashId $ \ hashIdPtr -> do
-                    allocBytesQuerying $ \ outPtr outLen -> botan_srp6_generate_verifier
-                        (ConstPtr identifierPtr)
-                        (ConstPtr passwordPtr)
-                        (ConstPtr saltPtr)
-                        saltLen
-                        (ConstPtr groupIdPtr)
-                        (ConstPtr hashIdPtr)
-                        outPtr
-                        outLen
+-- | Generate a new SRP-6 verifier
+srp6GenerateVerifier ::
+     Identifier      -- ^ __identifier__: a username or other client identifier
+  -> Password        -- ^ __password__: the secret used to authenticate user
+  -> Salt            -- ^ __salt[]__: a randomly chosen value, at least 128 bits long
+  -> DLGroupName     -- ^ __group_id__: specifies the shared SRP group
+  -> HashName        -- ^ __hash_id__: specifies a secure hash function
+  -> IO SRP6Verifier -- ^ __verifier[]__: out buffer to store the SRP-6 verifier value
+srp6GenerateVerifier identifier password salt groupId hashId =
+    asCString identifier $ \ identifierPtr ->
+    asCString password $ \ passwordPtr ->
+    asBytesLen salt $ \ saltPtr saltLen ->
+    asCString groupId $ \ groupIdPtr ->
+    asCString hashId $ \ hashIdPtr ->
+    createWithGroupSize groupId $ \ outPtr outLen ->
+    throwBotanIfNegative_ $
+      botan_srp6_generate_verifier
+        (ConstPtr identifierPtr)
+        (ConstPtr passwordPtr)
+        (ConstPtr saltPtr)
+        saltLen
+        (ConstPtr groupIdPtr)
+        (ConstPtr hashIdPtr)
+        outPtr
+        (castPtr outLen)
 
--- NOTE: ORDER IS DIFFERENT FROM SERVER GENERATE VERIFIER
--- | SRP6a Client side step 2:  Generate a client A-value and the client shared key
-srp6ClientAgree
-    :: Identifier   -- ^ __username__: the username we are attempting login for
-    -> Password     -- ^ __password__: the password we are attempting to use
-    -> DLGroupName  -- ^ __group_id__: specifies the shared SRP group
-    -> HashName     -- ^ __hash_id__: specifies a secure hash function
-    -> Salt         -- ^ __salt[]__: is the salt value sent by the server
-    -> SRP6BValue   -- ^ __uint8_t__: B[] is the server's public value
-    -> RNG          -- ^ __rng_obj__: is a random number generator object
-    -> IO (SRP6AValue, SRP6SharedSecret)    -- @(A,K)@
-srp6ClientAgree identifier password groupId hashId salt b rng = do
-    asCString identifier $ \ identifierPtr -> do
-        asCString password $ \ passwordPtr -> do
-            asCString groupId $ \ groupIdPtr -> do
-                asCString hashId $ \ hashIdPtr -> do
-                    asBytesLen salt $ \ saltPtr saltLen -> do
-                        asBytesLen b $ \ bPtr bLen -> do
-                            withRNG rng $ \ botanRNG -> do
-                                alloca $ \ aSzPtr -> do
-                                    alloca $ \ kSzPtr -> do
-                                        -- Query sizes
-                                        -- TODO: Actually ensure expected error (insufficient buffer space)
-                                        --  and propagate unexpected errors
-                                        _ <- botan_srp6_client_agree
-                                            (ConstPtr identifierPtr)
-                                            (ConstPtr passwordPtr)
-                                            (ConstPtr groupIdPtr)
-                                            (ConstPtr hashIdPtr)
-                                            (ConstPtr saltPtr)
-                                            saltLen
-                                            (ConstPtr bPtr)
-                                            bLen
-                                            botanRNG
-                                            nullPtr
-                                            aSzPtr
-                                            nullPtr
-                                            kSzPtr
-                                        kSz <- fromIntegral <$> peek kSzPtr
-                                        aSz <- fromIntegral <$> peek aSzPtr
-                                        allocBytesWith kSz $ \ kPtr -> do
-                                            allocBytes aSz $ \ aPtr -> do
-                                                throwBotanIfNegative_ $ botan_srp6_client_agree
-                                                    (ConstPtr identifierPtr)
-                                                    (ConstPtr passwordPtr)
-                                                    (ConstPtr groupIdPtr)
-                                                    (ConstPtr hashIdPtr)
-                                                    (ConstPtr saltPtr)
-                                                    saltLen
-                                                    (ConstPtr bPtr)
-                                                    bLen
-                                                    botanRNG
-                                                    aPtr
-                                                    aSzPtr
-                                                    kPtr
-                                                    kSzPtr
-
--- NOTE: Missing FFI function: srp6_group_identifierz
+-- | SRP6a Client side
+srp6ClientAgree ::
+     -- | __username__: the username we are attempting login for
+     Identifier
+     -- | __password__: the password we are attempting to use
+  -> Password
+     -- | __group_id__: specifies the shared SRP group
+  -> DLGroupName
+     -- | __hash_id__: specifies a secure hash function
+  -> HashName
+     -- | __salt[]__: is the salt value sent by the server
+  -> Salt
+     -- | __B[]__: is the server's public value
+  -> SRP6BValue
+     -- | __rng_obj__: is a random number generator object
+  -> RNG
+     -- | A tuple of two elements (in order):
+     --
+     -- * __A[]__: out buffer to store the SRP-6 A value
+     -- * __K[]__: out buffer to store the symmetric value
+  -> IO (SRP6AValue, SRP6SharedSecret)
+srp6ClientAgree identifier password groupId hashId salt b rng =
+    asCString identifier $ \ identifierPtr ->
+    asCString password $ \ passwordPtr ->
+    asCString groupId $ \ groupIdPtr ->
+    asCString hashId $ \ hashIdPtr ->
+    asBytesLen salt $ \ saltPtr saltLen ->
+    asBytesLen b $ \ bPtr bLen ->
+    withRNG rng $ \ botanRNG ->
+    createWithGroupSize' groupId $ \ aPtr aSzPtr ->
+    createWithGroupSize groupId $ \ kPtr kSzPtr ->
+    throwBotanIfNegative_ $
+      botan_srp6_client_agree
+        (ConstPtr identifierPtr)
+        (ConstPtr passwordPtr)
+        (ConstPtr groupIdPtr)
+        (ConstPtr hashIdPtr)
+        (ConstPtr saltPtr)
+        saltLen
+        (ConstPtr bPtr)
+        bLen
+        botanRNG
+        aPtr
+        aSzPtr
+        kPtr
+        kSzPtr
 
 -- | Return the size, in bytes, of the prime associated with group_id
-srp6GroupSize
-    :: DLGroupName  -- ^ __group_id__
-    -> IO Int       -- ^ __group_p_bytes__
-srp6GroupSize groupId = asCString groupId $ \ groupIdPtr -> do
+--
+-- This function can be used to determine the size of output buffers for
+-- generated keys in the SRP6 algorithm. Such buffers need to be allocated
+-- before calling SRP6 functions. An example of such a buffer is the
+-- @verifier[]@ buffer in the 'srp6GenerateVerifier' function. In @botan-low@,
+-- such buffers are created internally using 'srp6GroupSize', without requiring
+-- input from you (the user).
+srp6GroupSize ::
+     DLGroupName  -- ^ __group_id__
+  -> IO Int       -- ^ __group_p_bytes__
+srp6GroupSize groupId =
+    asCString groupId $ \ groupIdPtr ->
     alloca $ \ szPtr -> do
-        throwBotanIfNegative_ $ botan_srp6_group_size (ConstPtr groupIdPtr) szPtr
-        fromIntegral <$> peek szPtr
+      throwBotanIfNegative_ $
+        botan_srp6_group_size (ConstPtr groupIdPtr) szPtr
+      fromIntegral <$> peek szPtr
+
+{-------------------------------------------------------------------------------
+  Utility
+-------------------------------------------------------------------------------}
+
+-- | A version of 'BS.create' that determines the size of the byte string based
+-- on an argument 'DLGroupName'.
+createWithGroupSize ::
+     DLGroupName
+  -> (Ptr Word8 -> Ptr CSize -> IO ())
+  -> IO ByteString
+createWithGroupSize groupId k = do
+    sz <- srp6GroupSize groupId
+    createWithSize sz k
+
+-- | A version of 'BS.create' that also creates a pointer for the size of the
+-- byte string.
+createWithSize ::
+     Int
+  -> (Ptr Word8 -> Ptr CSize -> IO ())
+  -> IO ByteString
+createWithSize sz k =
+    BS.createUptoN sz $ \bytesPtr ->
+    alloca $ \lenPtr -> do
+      poke lenPtr (fromIntegral sz)
+      k bytesPtr lenPtr
+      fromIntegral <$> peek lenPtr
+
+-- | A version of 'BS.createUptoN'' that determines the size of the byte string
+-- based on an argument 'DLGroupName'.
+createWithGroupSize' ::
+     DLGroupName
+  -> (Ptr Word8 -> Ptr CSize -> IO a)
+  -> IO (ByteString, a)
+createWithGroupSize' groupId k = do
+    sz <- srp6GroupSize groupId
+    createWithSize' sz k
+
+-- | A version of 'BS.createUptoN'' that also creates a pointer for the size of
+-- the byte string.
+createWithSize' ::
+     Int
+  -> (Ptr Word8 -> Ptr CSize -> IO a)
+  -> IO (ByteString, a)
+createWithSize' sz k =
+    BS.createUptoN' sz $ \bytesPtr ->
+    alloca $ \lenPtr -> do
+      poke lenPtr (fromIntegral sz)
+      x <- k bytesPtr lenPtr
+      sz' <- fromIntegral <$> peek lenPtr
+      pure (sz', x)